avoid integer overflow if running for more than ~18h20min, avoid division by zero

+++ lambda/scheduler.h
@@ -29,7 +29,7 @@
 typedef struct {
 	void (*func)(void);
 	uint16_t delay;
-	uint16_t time;
+	uint32_t time;
 	bool done;
 } Task;
